Update weather.handlebars
The button for units change is in the right panel now. 
Change in weather forecast display. I will work on that more later as now everything looks great when we are retrieving a weather forecast for one city only. If there are more than one city - there's a total mess in weather forecast display.